Prefer source collection over HI files
The interface metadata collection does not handle type classes well.
It finds the instances of reexported type class functions that
are included in the hi files, but this list can be confusing and
does not merge well.  For instance in Data.List it finds

null :: [a] -> Bool and null :: PSQ a -> Bool but the the source
collector finds null :: Foldable t => t a -> Bool.  The interface
collector does not find the general type and the merge code does not
work well.