Fix leksah-metadata.nix leksah-server ver mismatch
The trick of using nix-build to build leksah metadata by adding
leksah-server to the projects default.nix does not always work
and is fragile because the wrong version of leksah-server might be used.

This change pulls a particular commit of leksah-server from github
so that it does not need to be added to the project manually and the
right version will be used.

Unfortunately the commit rev and sha256 will need to be updated manually
in the leksah-metadata.nix file.